Fix undefined method error
The Hash#to_params method is not supplied by Rails (where the
Hash#to_param) method is. It seems that one of the gems we used to use
in our app provided that method, which we no longer use. The
Rails-supplied Hash#to_param has the same output, so switch to using
this.